fix: 🐛 import Framer Motion's components on client-side
